% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_cluster.R
\name{bayes_cluster}
\alias{bayes_cluster}
\title{Title}
\usage{
bayes_cluster(y, E, population, sp.obj, centroids, max.prop, shape, rate, J,
  pi0, n.sim.lambda, n.sim.prior, n.sim_post, burnin.prop = 0.1,
  theta.init = vector(mode = "numeric", length = 0))
}
\arguments{
\item{y}{}

\item{E}{}

\item{population}{}

\item{sp.obj}{}

\item{centroids}{}

\item{max.prop}{}

\item{shape}{}

\item{rate}{}

\item{J}{}

\item{pi0}{}

\item{burnin.prop}{}

\item{theta.init}{}

\item{n_sim.lambda}{}

\item{n_sim.prior}{}

\item{n_sim_post}{}
}
\description{
Title
}

