% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besag_newell.R
\name{besag_newell}
\alias{besag_newell}
\title{Title}
\usage{
besag_newell(geo, population, cases, expected.cases = NULL, k, alpha.level)
}
\arguments{
\item{geo}{}

\item{population}{}

\item{cases}{}

\item{expected.cases}{}

\item{k}{}

\item{alpha.level}{}
}
\value{

}
\description{
Title
}
