\name{besagNewell}
\alias{besagNewell}
\title{Internal Besag Newell method}
\description{Internal}
\usage{
besagNewell(observedCases, expectedCases, nearestNeighborsList, nZones, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{observedCases}{observed cases for each area}
  \item{expectedCases}{Expected cases for each area}
  \item{nearestNeighbors}{See \code{\link{create_geo_objects}}}
  \item{nZones}{number of single zones}
  \item{k}{max number of cases in each}
}
%\value{}
\author{Albert Y. Kim}
\seealso{\code{\link{besag_newell}}}
\keyword{internal}
