% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeAllLogLkhd}
\alias{computeAllLogLkhd}
\title{Compute log likelihood ratio test statistic for all single zones.}
\usage{
computeAllLogLkhd(observedCases, expectedCases, nearestNeighborsList,
  logLkhdType)
}
\arguments{
\item{observedCases}{}

\item{expectedCases}{}

\item{nearestNeighborsList}{}

\item{logLkhdType}{}

\item{nZones}{}
}
\value{
Vector of log likelihood
}
\description{
Compute log likelihood ratio test statistic for all single zones.
}

