% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geo_objects.R
\name{create_geo_objects}
\alias{create_geo_objects}
\title{Create geographical objects to be used in Bayesian Cluster Detection Method}
\usage{
create_geo_objects(max.prop, population, centroids, spatial.obj)
}
\arguments{
\item{max.prop}{maximum proportion of study region's population each single zone can contain}

\item{population}{vector of length \code{n} of the population of each area}

\item{centroids}{\verb{n x 2} table of the (x,y)-coordinates of the area centroids.  The coordinate system must be grid-based}

\item{spatial.obj}{an sf object representing the study region}
}
\value{
\item{overlap}{list with two elements: \verb{1. presence} which lists for each area all the single zones it is present in and \verb{2. cluster.list} for each single zone its component areas}
\item{cluster.coords}{\verb{n.zones x 2} matrix of the center and radial area of each single zone}
}
\description{
This internal function creates the geographical objects needed to run the Bayesian cluster detection method in \code{\link[=bayes_cluster]{bayes_cluster()}}.  Specifically it creates all single zones based data objects, where single zones are the \emph{zones} defined by Kulldorff (1997).
}
\examples{
data(pennLC_sf)
# Disable s2 spherical geometry for centroid calculation
sf::sf_use_s2(FALSE)
max.prop <- 0.15
# Aggregate population by county (first 67 rows are unique counties)
population <- tapply(pennLC_sf$population, pennLC_sf$county, sum)
# Get centroids
centroids <- sf::st_coordinates(sf::st_centroid(
  sf::st_as_sf(aggregate(pennLC_sf["geometry"], by = list(pennLC_sf$county), FUN = head, n = 1))
))
centroids <- latlong2grid(centroids)
# Get unique county geometries
county_sf <- sf::st_as_sf(aggregate(pennLC_sf["geometry"],
  by = list(county = pennLC_sf$county), FUN = head, n = 1))
output <- create_geo_objects(max.prop, population, centroids, county_sf)
## number of single zones
nrow(output$cluster.coords)
}
\references{
Wakefield J. and Kim A.Y. (2013) A Bayesian model for cluster detection.\emph{Biostatistics}, \strong{14}, 752--765.
}
\author{
Albert Y. Kim
}
