% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_objects.R
\name{define_single_zones}
\alias{define_single_zones}
\title{Define a study region's single zones.}
\usage{
define_single_zones(centroids, population, pop_upper_bound)
}
\arguments{
\item{centroids}{data frame with 2 columns of the centroids of the n areas 
in the study region.}

\item{population}{vector of length n of the corresponding population 
counts.}

\item{pop_upper_bound}{upper bound of the proportion of the study 
region's population the single zones can contain.}
}
\value{
A list with two objects
\item{nearest_neighbors}{A list of length n of the single zones for each area} 
\item{cluster_coords}{A data frame with 2 columns of the centering 
and radial area for each single zone.}
}
\description{
Using the centroids of the n areas in a study region, define the set of 
single zones based on the upper bound \code{pop_upper_bound} of the 
proportion of the study region's population each single zone can contain.
}
\examples{
data(NYleukemia)
centroids <- sp::coordinates(NYleukemia$spatial.polygon)
single_zones <- define_single_zones(centroids, NYleukemia$data$population, 0.15)
}
\references{
Kulldorff, M. (1997) A spatial scan statistic. 
\emph{Communications in Statistics: Theory and Methods}, \bold{26}, 1481--1496.
}
\seealso{
\code{\link{create_geo_objects}}
}

