% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaPriorCh.R
\name{GammaPriorCh}
\alias{GammaPriorCh}
\title{Compute Parameters to Calibrate a Gamma Distribution}
\usage{
GammaPriorCh(theta, prob, d)
}
\arguments{
\item{theta}{}

\item{prob}{}

\item{d}{}
}
\value{

}
\description{
Compute Parameters to Calibrate a Gamma Distribution
}
