% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_cluster.R
\name{kulldorff}
\alias{kulldorff}
\title{Title}
\usage{
kulldorff(geo, cases, population, expected.cases = NULL, pop.upper.bound,
  n.simulations, alpha.level, plot = TRUE)
}
\arguments{
\item{geo}{}

\item{cases}{}

\item{population}{}

\item{expected.cases}{}

\item{pop.upper.bound}{}

\item{n.simulations}{}

\item{alpha.level}{}

\item{plot}{}
}
\description{
Title
}

