% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2grid.R
\name{latlong2grid}
\alias{latlong2grid}
\title{Convert Coordinates from Latitude/Longitude to Grid}
\usage{
latlong2grid(input)
}
\arguments{
\item{input}{either an \verb{n x 2} matrix of longitude and latitude coordinates in decimal format or an sf/sfc object}
}
\value{
Either a data frame with the corresponding (x,y) kilometer-based grid coordinates, or an sf/sfc object with the coordinates transformed.
}
\description{
Convert geographic latitude/longitude coordinates to kilometer-based grid coordinates.
}
\details{
Longitude/latitudes are not a grid-based coordinate system:  latitudes are equidistant but the distance between longitudes varies.
}
\note{
Rough conversion of US lat/long to km (used by GeoBUGS):  (see
also forum.swarthmore.edu/dr.math/problems/longandlat.html).
Radius of earth: r = 3963.34 (equatorial) or 3949.99 (polar) mi =
6378.2 or 6356.7 km, which implies: km per mile  = 1.609299 or
1.609295 a change of 1 degree of latitude corresponds to the same
number of km, regardless of longitude.  arclength=r*theta, so the
multiplier for coord y should probably be just the radius of
earth. On the other hand, a change of 1 degree in longitude
corresponds to a different distance, depending on latitude.  (at N
pole, the change is essentially 0.  at the equator, use equatorial
radius.
}
\examples{
## Convert coordinates
coord <- data.frame(rbind(
 # Montreal, QC:  Latitude: 45deg 28' 0" N (deg min sec), Longitude: 73deg 45' 0" W
 c(-73.7500, 45.4667),
 # Vancouver, BC:  Latitude: 45deg 39' 38" N (deg min sec), Longitude: 122deg 36' 15" W
 c(-122.6042, 45.6605)
))
latlong2grid(coord)

## Convert sf object
data(pennLC_sf)
# Get unique county geometries
county_sf <- sf::st_as_sf(aggregate(pennLC_sf["geometry"],
  by = list(county = pennLC_sf$county), FUN = head, n = 1))
new <- latlong2grid(county_sf)
par(mfrow=c(1,2))
plot(sf::st_geometry(county_sf), axes=TRUE)
title("Lat/Long")
plot(sf::st_geometry(new), axes=TRUE)
title("Grid (in km)")
}
\author{
Lance A. Waller
}
