% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leglabs.R
\name{leglabs}
\alias{leglabs}
\title{Make legend labels}
\usage{
leglabs(vec, under = "under", over = "over", between = "-", reverse = FALSE)
}
\arguments{
\item{vec}{vector of break values}

\item{under}{character value for under}

\item{over}{character value for over}

\item{between}{character value for between}

\item{reverse}{flag to reverse order of values, you will also need to reorder colours, see example}
}
\description{
leglabs makes character strings from the same break points. This function was copied from the soon-to-be
deprecated \code{maptools} package with permission from author Roger Bivand
}
\author{
Roger Bivand, Nick Bearman, Nicholas Lewin-Koh
}
