% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_cluster.R
\name{LogNormalPriorCh}
\alias{LogNormalPriorCh}
\title{Title}
\usage{
LogNormalPriorCh(theta1, theta2, prob1, prob2)
}
\arguments{
\item{theta1}{}

\item{theta2}{}

\item{prob1}{}

\item{prob2}{}
}
\description{
Title
}

