% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kulldorff.R
\name{map_kulldorff}
\alias{map_kulldorff}
\title{Plot results of \code{kulldorff} method}
\usage{
map_kulldorff(sp_obj, kulldorff_output, ...)
}
\arguments{
\item{sp_obj}{SpatialPolygonsDataFrame object}

\item{kulldorff_output}{List output of \code{\link{kulldorff}}}

\item{...}{other arguments passed on to  \code{\link{map_variable}}}
}
\value{
a ggplot object
}
\description{
Wrapper function for \code{\link{map_variable}} to plot the results of a kulldorff analysis.
}
\examples{
data("NYleukemia")
centroids <- sp::coordinates(NYleukemia)
output <- kulldorff(centroids = centroids,
   cases = NYleukemia$cases, population = NYleukemia$population,
   expected_cases = NULL, pop_upper_bound = 0.15, alpha_level=0.05,
   n_sim=999)
map_kulldorff(NYleukemia, output)
}

