% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_variable}
\alias{map_variable}
\title{Plot ggplot compliant map}
\usage{
map_variable(sp_obj, variable = NULL, x_label = "eastings",
  y_label = "northings", title = NULL, legend_label = NULL)
}
\arguments{
\item{sp_obj}{SpatialPolygonsDataFrame object}

\item{variable}{Variable within data frame \code{sp_obj@data}}

\item{x_label}{Label for eastings}

\item{y_label}{Label for northings}

\item{title}{Plot title}

\item{legend_label}{label for legend. Defaults to variable name.}
}
\value{
a ggplot object
}
\description{
\url{https://github.com/hadley/ggplot2/wiki/plotting-polygon-shapefiles}
}
\examples{
shapefile_name <- system.file("extdata/sids/sids.shp", package="SpatialEpi")
proj4string <- sp::CRS("+proj=longlat +ellps=clrk66")
nc_sids <- maptools::readShapePoly(
   fn = shapefile_name,
   ID = "FIPSNO",
   proj4string = proj4string
   )
names(nc_sids)
nc_sids$RATE74 <- nc_sids$SID74/nc_sids$BIR74
nc_map <- map_variable(nc_sids, "RATE74")
nc_map
}

