% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvariable.R
\name{mapvariable}
\alias{mapvariable}
\title{Plot Levels of a Variable in a Colour-Coded Map}
\usage{
mapvariable(
  data,
  y,
  ncut = 1000,
  nlevels = 10,
  lower = NULL,
  upper = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{an sf object containing the variable to plot}

\item{y}{column name (unquoted) or character string specifying the variable to plot.
Must be a column in \code{data}.}

\item{ncut}{number of cuts in colour levels to plot}

\item{nlevels}{number of levels to include in legend}

\item{lower}{lower bound of levels}

\item{upper}{upper bound of levels}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}
}
\value{
A map colour-coded to indicate the different levels of \code{y}. Returns \code{data} invisibly
for use in pipelines.
}
\description{
Plot levels of a variable in a colour-coded map along with a legend.
}
\examples{
data(scotland_sf)

# Using pipe syntax (recommended)
scotland_sf |> mapvariable(cases)

# With transformations
scotland_sf$SMR <- scotland_sf$cases / scotland_sf$expected
scotland_sf |> mapvariable(SMR, main="Scotland SMR")

# Using character column name
scotland_sf |> mapvariable("cases", main="Scotland", xlab="Eastings (km)", ylab="Northings (km)")
}
\references{
Pebesma, E. (2018) Simple Features for R: Standardized Support for Spatial Vector Data. \emph{The R Journal}, \strong{10}, 439--446.
}
\author{
Jon Wakefield, Nicky Best, Sebastien Haneuse, and Albert Y. Kim
}
