% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon2sf.R
\name{polygon2sf}
\alias{polygon2sf}
\title{Convert a Polygon to an sf Object}
\usage{
polygon2sf(poly, crs, area.names = NULL, nrepeats = NULL)
}
\arguments{
\item{poly}{a 2-column matrix of coordinates, where each complete subpolygon is separated by NA's}

\item{crs}{the coordinate reference system, either as EPSG code (e.g., 4326) or proj4 string}

\item{area.names}{names of all areas}

\item{nrepeats}{number of sub polygons for each area}
}
\value{
An object of class \code{sf} with POLYGON or MULTIPOLYGON geometries.
}
\description{
Converts a polygon (a matrix of coordinates with NA values to separate subpolygons) into an sf object.
}
\details{
Just as when plotting with the \code{\link[graphics:polygon]{graphics::polygon()}} function, it is assumed that each subpolygon is to be closed by joining the last point to the first point. In the matrix \code{poly}, NA values separate complete subpolygons.
In the case with an area consists of more than one separate closed polygon, \code{nrepeats} specifies the number of closed polygons associated with each area.
}
\examples{
data(scotland)

polygon <- scotland$polygon$polygon
names <- scotland$data$county.names
nrepeats <- scotland$polygon$nrepeats

# Create sf object with WGS84 CRS
sf_polygon <- polygon2sf(polygon, crs = 4326, names, nrepeats)

par(mfrow=c(1,2))
# plot using polygon function
plot(polygon, type='n', xlab="Eastings (km)", ylab="Northings (km)", main="Polygon File")
polygon(polygon)

# plot as sf object
plot(sf::st_geometry(sf_polygon), axes=TRUE)
title(xlab="Eastings (km)", ylab="Northings (km)", main="SF Object")

# Note that area 23 (argyll-bute) consists of 8 separate polygons
nrepeats[23]
plot(sf::st_geometry(sf_polygon[23, ]), add=TRUE, col="red")
}
\references{
Pebesma, E. (2018) Simple Features for R: Standardized Support for Spatial Vector Data. \emph{The R Journal}, \strong{10}, 439--446.
}
\author{
Albert Y. Kim
}
