% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_objects.R
\name{polygon2spatial_polygon}
\alias{polygon2spatial_polygon}
\title{Title}
\usage{
polygon2spatial_polygon(poly, coordinate.system, area.names = NULL,
  nrepeats = NULL)
}
\arguments{
\item{poly}{}

\item{coordinate.system}{}

\item{area.names}{}

\item{nrepeats}{}
}
\description{
Title
}

