% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kulldorff.R
\name{return_single_zone_areas}
\alias{return_single_zone_areas}
\title{Obtain all areas included in a single zone}
\usage{
return_single_zone_areas(cluster_index, zone_info)
}
\arguments{
\item{cluster_index}{integer index of single zone}

\item{zone_info}{output of \code{\link{define_single_zones}} function}
}
\value{
vector of areas
}
\description{
Obtain all areas included in a single zone
}
\examples{
1+1
}

