% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotland.R
\docType{data}
\name{scotland}
\alias{scotland}
\title{Lip Cancer in Scotland}
\format{
List containing:
\describe{
\item{geo}{a table of county IDs, x-coordinates (eastings) and y-coordinates (northings) of the geographic centroid of each county.}
\item{data}{a table of county IDs, number of cases, population and strata information}
\item{spatial.polygon}{a Spatial Polygons class map of Scotland (deprecated, use scotland_sf instead)}
\item{polygon}{a polygon map of Scotland (see \code{\link[=polygon2sf]{polygon2sf()}})}
}
}
\source{
Kemp I., Boyle P., Smans M. and Muir C. (1985) Atlas of cancer in Scotland, 1975-1980, incidence and epidemiologic perspective \emph{International Agency for Research on Cancer} \strong{72}.
}
\usage{
scotland
}
\description{
County-level (n=56) data for lip cancer among males in Scotland between 1975-1980
}
\examples{
# Use scotland_sf for modern sf-based workflows
data(scotland_sf)
SMR <- scotland_sf$cases / scotland_sf$expected
mapvariable(SMR, scotland_sf)





}
\references{
Clayton D. and Kaldor J. (1987) Empirical Bayes estimates of age-standardized relative risks for use in disease mapping.  \emph{Biometrics}, \strong{43}, 671--681.
}
\keyword{datasets}
