% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotland_sf.R
\docType{data}
\name{scotland_sf}
\alias{scotland_sf}
\title{Scotland SF Object}
\format{
A data frame with 56 rows representing counties and 5 variables:
\describe{
  \item{geometry}{Scotland counties}
  \item{cases}{indicates that regional subsidiaries are included}
  \item{county.names}{available seat kilometers flown every week}
  \item{AFF}{available seat kilometers flown every week}
  \item{expected}{available seat kilometers flown every week}
}
}
\source{
Kemp I., Boyle P., Smans M. and Muir C. (1985) Atlas of cancer in Scotland, 1975-1980, incidence and epidemiologic perspective \emph{International Agency for Research on Cancer} \bold{72}.
}
\usage{
scotland_sf
}
\description{
scotland_sf is a spatial object that can be used to create static or interactive maps for lip cancer among males in Scotland between 1975-1980.
}
\keyword{datasets}
