% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialEpi.R
\docType{package}
\name{SpatialEpi}
\alias{SpatialEpi}
\alias{SpatialEpi-package}
\title{SpatialEpi: Data and methods for spatial epidemiology.}
\description{
This package provides methods and Data for Spatial Epidemiology.
}
\examples{
# Example usage
library(SpatialEpi)
data(scotland)
map <- scotland$spatial.polygon
y <- scotland$data$cases
E <- scotland$data$expected
SMR <- y/E
# Plot SMR
plotmap(SMR, map, nclr=9, location="topleft") 
}
