% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binomialLogLkhd}
\alias{binomialLogLkhd}
\title{Compute binomial likelihood ratio test statistic for Kulldorff method.}
\usage{
binomialLogLkhd(cz, nz, C, N)
}
\arguments{
\item{cz}{Disease count inside single zone.}

\item{nz}{Expected count inside single zone.}

\item{C}{Total disease count in region.}

\item{N}{Total expected count in region.}
}
\value{
log of likelihood ratio test statistic.
}
\description{
Compute binomial likelihood ratio test statistic for Kulldorff method.
}

