% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeAllLogLkhd}
\alias{computeAllLogLkhd}
\title{Compute log likelihood ratio test statistic for all single zones.}
\usage{
computeAllLogLkhd(observedCases, expectedCases, nearestNeighborsList,
  logLkhdType)
}
\arguments{
\item{observedCases}{vector of observed cases}

\item{expectedCases}{vector of expected cases}

\item{nearestNeighborsList}{list of nearest neighbors}

\item{logLkhdType}{type of log-likelihood}
}
\value{
Vector of log likelihoods for each single zone
}
\description{
Compute log likelihood ratio test statistic for all single zones.
}

