% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_objects.R
\name{create_geo_objects}
\alias{create_geo_objects}
\title{Create geographic objects needed for Bayesian method}
\usage{
create_geo_objects(centroids, population, pop_upper_bound)
}
\arguments{
\item{centroids}{data frame with 2 columns of the centroids of the n areas 
in the study region.}

\item{population}{vector of length n of the corresponding population 
counts.}

\item{pop_upper_bound}{upper bound of the proportion of the study 
region's population the single zones can contain.}
}
\value{
A list with three objects
\item{presence}{List of length n, for each area, the single zones it is present in.}
\item{cluster_list}{List of length n_zones indicating the areas included in each single zone}
\item{cluster_coords}{A data frame with 2 columns of the centering and radial area for each single zone.}
}
\description{
Create geographic objects needed for Bayesian method
}
\examples{
data(NYleukemia)
centroids <- sp::coordinates(NYleukemia)
geo_objects <- create_geo_objects(centroids, NYleukemia$population, 0.15)
}
\seealso{
\code{\link{define_single_zones}}
}
