% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_cluster.R
\name{estimate_lambda}
\alias{estimate_lambda}
\title{Title}
\usage{
estimate_lambda(n_sim, J, prior_z, overlap, pi0)
}
\arguments{
\item{n_sim}{}

\item{J}{}

\item{prior_z}{}

\item{overlap}{}

\item{pi0}{}
}
\description{
Title
}

