% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kulldorff.R
\name{get_cluster_report}
\alias{get_cluster_report}
\title{Generate cluster reports}
\usage{
get_cluster_report(cluster_index, population, cases, expected_cases, zone_info,
  all_log_lkhd, lambdas)
}
\arguments{
\item{cluster_index}{}

\item{population}{}

\item{cases}{}

\item{expected_cases}{}

\item{zone_info}{}

\item{all_log_lkhd}{}

\item{lambdas}{}
}
\value{
Cluster report
}
\description{
Return a report on most likely cluster and secondary clusters containing the
same fields as SatScan
}
\examples{
1+1
}

