% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kulldorff.R
\name{kulldorff}
\alias{kulldorff}
\title{Kulldorff method for cluster detection}
\usage{
kulldorff(centroids, cases, population, expected_cases = NULL,
  pop_upper_bound, alpha_level, n_sim = 9999, plot = FALSE)
}
\arguments{
\item{centroids}{data frame with 2 columns of the centroids of the n areas
in the study region.}

\item{cases}{Vector of number of cases for each area}

\item{population}{vector of length n of the corresponding population
counts.}

\item{expected_cases}{Vector of number of expected cases for each area}

\item{pop_upper_bound}{upper bound of the proportion of the study
  region's population the single zones can contain.}

\item{alpha_level}{alpha-level threshold used to declare significance}

\item{n_sim}{number of Monte Carlo samples used for significance measures}

\item{plot}{logical variable of whether to plot Monte Carlo distribution}
}
\value{
A list containing \describe{
  \item{\code{most_likely_cluster}}{\code{\link{get_cluster_report}} report on most likely cluster}
  \item{\code{secondary_clusters}}{list of \code{\link{get_cluster_report}} reports on secondary clusters; if none \code{NULL} is returned}
  \item{\code{type}}{type of likelihood: binomial or poisson}
  \item{\code{log_lkhd}}{log-likelihood of each zone considered}
  \item{\code{simulated_log_lkhd}}{\code{n.simulations} Monte Carlo samples of the log-likelihood of the most likely cluster}
  }
}
\description{
Kulldorff method for cluster detection
}
\examples{
data("NYleukemia")
centroids <- sp::coordinates(NYleukemia)
output <- kulldorff(centroids = centroids,
   cases = NYleukemia$cases, population = NYleukemia$population,
   expected_cases = NULL, pop_upper_bound = 0.15, alpha_level=0.05,
   n_sim=999)
output$most_likely_cluster
}

