% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kulldorffMC}
\alias{kulldorffMC}
\title{Compute log likelihood ratio test statistic for all}
\usage{
kulldorffMC(permutedCaseMatrix, expectedCases, nearestNeighborsList,
  logLkhdType)
}
\arguments{
\item{permutedCaseMatrix}{matrix of null simulated cases}

\item{expectedCases}{vector of expected cases}

\item{nearestNeighborsList}{list of nearest neighbors}

\item{logLkhdType}{type of log-likelihood}
}
\value{
Max
}
\description{
Compute log likelihood ratio test statistic for all
}

