% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kulldorffMC}
\alias{kulldorffMC}
\title{Compute log likelihood ratio test statistic for all}
\usage{
kulldorffMC(permutedCaseMatrix, expectedCases, nearestNeighborsList,
  logLkhdType)
}
\arguments{
\item{permutedCaseMatrix}{blah}

\item{expectedCases}{}

\item{nearestNeighborsList}{}

\item{logLkhdType}{}
}
\value{
Max
}
\description{
Compute log likelihood ratio test statistic for all
}

