% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapvariable}
\alias{mapvariable}
\title{Title}
\usage{
mapvariable(y, spatial.polygon, ncut = 1000, nlevels = 10, lower = NULL,
  upper = NULL, main = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{y}{}

\item{spatial.polygon}{}

\item{ncut}{}

\item{nlevels}{}

\item{lower}{}

\item{upper}{}

\item{main}{}

\item{xlab}{}

\item{ylab}{}
}
\description{
Title
}

