% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvariable.R
\name{mapvariable}
\alias{mapvariable}
\title{Plot Levels of a Variable in a Colour-Coded Map}
\usage{
mapvariable(
  y,
  spatial.polygon,
  ncut = 1000,
  nlevels = 10,
  lower = NULL,
  upper = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{y}{variable to plot}

\item{spatial.polygon}{an sf object}

\item{ncut}{number of cuts in colour levels to plot}

\item{nlevels}{number of levels to include in legend}

\item{lower}{lower bound of levels}

\item{upper}{upper bound of levels}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}
}
\value{
A map colour-coded to indicate the different levels of \code{y}
}
\description{
Plot levels of a variable in a colour-coded map along with a legend.
}
\examples{
data(scotland_sf)
y <- scotland_sf$cases
E <- scotland_sf$expected
SMR <- y/E
mapvariable(SMR, scotland_sf, main="Scotland", xlab="Eastings (km)", ylab="Northings (km)")
}
\references{
Pebesma, E. (2018) Simple Features for R: Standardized Support for Spatial Vector Data. \emph{The R Journal}, \strong{10}, 439--446.
}
\author{
Jon Wakefield, Nicky Best, Sebastien Haneuse, and Albert Y. Kim
}
