% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NYleukemia}
\alias{NYleukemia}
\title{Leukemia in Upstate New York}
\format{A SpatialPolygonsDataFrame object}
\usage{
data(NYleukemia)
}
\description{
Get data
}
\details{
\itemize{
  \item price. price in US dollars (\$326--\$18,823)
  \item carat. weight of the diamond (0.2--5.01)
}
}
\keyword{datasets}

