% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NYleukemia}
\alias{NYleukemia}
\title{Leukemia in Upstate New York.}
\format{object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}} with a data frame of
281 observations and 4 variables:
\itemize{
  \item ID: id variable
  \item FIPS: Federal Information Processing Standards code of census tract
  \item population: population size (1980 U.S. Census)
  \item cases: number of cases 1978-1982
}}
\source{
\url{http://web1.sph.emory.edu/users/lwaller/ch9index.htm}
}
\usage{
NYleukemia
}
\description{
Leukemia data for 281 census tracts in 7 counties in Upstate New York.
}
\references{
Waller, L. and C. Gotway (2004) Applied Spatial Statistics for
Public Health Data. New York: John Wiley and Sons.
}
\keyword{datasets}
