% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pennLC}
\alias{pennLC}
\title{Pennsylvania Lung Cancer.}
\format{object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}} with a data frame of
281 observations and 4 variables:
\itemize{
  \item ID: id variable
  \item FIPS: Federal Information Processing Standards code of census tract
  \item population: population size (1980 U.S. Census)
  \item cases: number of cases 1978-1982
}}
\source{
Population data was obtained from the 2000 decennial census, lung 
cancer and smoking data were obtained from the Pennsylvania Department of 
Health website: \url{http://www.dsf.health.state.pa.us/}
}
\usage{
pennLC
}
\description{
County-level (n=67) population/case data for lung cancer in Pennsylvania in 
2002, stratified on race (white vs non-white), gender and age (Under 40, 
40-59, 60-69 and 70+).  Additionally, county-specific smoking rates.
}
\keyword{datasets}

