% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_objects.R
\name{pennLC_sf}
\alias{pennLC_sf}
\title{Pennsylvania Lung Cancer SF Object}
\value{
\code{pennLC_sf}
}
\description{
pennLC_sf is a spatial object that can be used to create static or interactive maps.
}
\examples{
Static Map

ggplot() + 
geom_sf(data = pennLC_sf)

Interactive Map using Leaflet

leaflet(pennLC_sf) \%>\%
addPolygons()

}
