% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_objects.R
\docType{data}
\name{pennLC_sf}
\alias{pennLC_sf}
\title{Pennsylvania Lung Cancer SF Object}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1072 rows and 7 columns.
}
\usage{
pennLC_sf
}
\description{
pennLC_sf is a spatial object that is used to create static or interactive maps.
}
\examples{
Static Map

ggplot() + 
geom_sf(data = pennLC_sf)

Interactive Map using Leaflet

leaflet(pennLC_sf) \%>\%
addPolygons()

}
\keyword{datasets}
