% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{plotmap}
\alias{plotmap}
\title{Title}
\usage{
plotmap(values, map, log = FALSE, nclr = 7, include.legend = TRUE,
  lwd = 0.5, round = 3, brks = NULL, legend = NULL,
  location = "topright", rev = FALSE, leg.cex = 1)
}
\arguments{
\item{values}{}

\item{map}{}

\item{log}{}

\item{nclr}{}

\item{include.legend}{}

\item{lwd}{}

\item{round}{}

\item{brks}{}

\item{legend}{}

\item{location}{}

\item{rev}{}

\item{leg.cex}{}
}
\description{
Title
}
\examples{

}

