% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_cluster.R
\name{process_MCMC_sample}
\alias{process_MCMC_sample}
\title{Title}
\usage{
process_MCMC_sample(sample, param, RR_area, cluster_list, cutoffs)
}
\arguments{
\item{sample}{}

\item{param}{}

\item{RR_area}{}

\item{cluster_list}{}

\item{cutoffs}{}
}
\description{
Title
}

