% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_objects.R
\name{return_single_zone_areas}
\alias{return_single_zone_areas}
\title{Obtain all areas included in a single zone}
\usage{
return_single_zone_areas(single_zone_index, zone_info)
}
\arguments{
\item{single_zone_index}{integer index of single zone}

\item{zone_info}{output of \code{\link{define_single_zones}} defining single
zones}
}
\value{
vector of areas
}
\description{
Obtain all areas included in a single zone
}
\examples{
1+1
}
