% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland}
\alias{scotland}
\title{Lip Cancer in Scotland.}
\format{object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}} with a data frame of
281 observations and 4 variables:
\itemize{
  \item ID: id variable
  \item FIPS: Federal Information Processing Standards code of census tract
  \item population: population size (1980 U.S. Census)
  \item cases: number of cases 1978-1982
}}
\source{
Kemp I., Boyle P., Smans M. and Muir C. (1985) Atlas of cancer in 
Scotland, 1975-1980, incidence and epidemiologic perspective 
\emph{International Agency for Research on Cancer} \bold{72}.
}
\usage{
scotland
}
\description{
County-level (n=56) data for lip cancer among males in Scotland between 
1975-1980.
}
\references{
Clayton D. and Kaldor J. (1987) Empirical Bayes estimates of 
age-standardized relative risks for use in disease mapping.
\emph{Biometrics}, \bold{43}, 671--681
}
\keyword{datasets}
