% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialEpi.R
\docType{package}
\name{SpatialEpi}
\alias{SpatialEpi}
\alias{SpatialEpi-package}
\title{SpatialEpi: A package for methods and data for spatial epidemiology.}
\description{
The SpatialEpi package provides three categories of important functions:
SpatialEpi, bar and baz.
}
\section{SpatialEpi functions}{

The SpatialEpi functions ...
}

