% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialEpi.R
\docType{package}
\name{SpatialEpi}
\alias{SpatialEpi}
\alias{SpatialEpi-package}
\title{SpatialEpi: Data and methods for spatial epidemiology.}
\description{
This package provides...
}
\section{Foo functions}{

The foo functions ...
}

